oreder_micro_service/requirements.txt
Flask==3.1.0
kafka-python>=2.0.2